Remove TransactionStore and SQLiteTransactionStore
Refactor completed with Claude Code
