+++ src/bin/server.rs
@@ -3,7 +3,7 @@ use std::{
     fs::OpenOptions,
     net::{Ipv4Addr, SocketAddr},
     process::exit,
-    sync::{Arc, Mutex},
+    sync::Arc,
 };
 
 use axum::{
@@ -20,10 +20,7 @@ use tower_livereload::LiveReloadLayer;
 
 use tracing_subscriber::{Layer, filter, layer::SubscriberExt, util::SubscriberInitExt};
 
-use budgeteur_rs::{
-    AppState, build_router, db::initialize, graceful_shutdown, logging_middleware,
-    stores::sqlite::SQLiteTransactionStore,
-};
+use budgeteur_rs::{AppState, build_router, graceful_shutdown, logging_middleware};
 
 /// The REST API server for budgeteur_rs.
 #[derive(Parser, Debug)]
@@ -73,17 +70,13 @@ async fn main() {
         "Could not open database file at {}: ",
         args.db_path
     ));
-    if let Err(error) = initialize(&conn) {
-        eprintln!("Could not initialize database: {error}");
-        exit(1);
-    }
-    let conn = Arc::new(Mutex::new(conn));
-    let app_config = AppState::new(
-        &secret,
-        Default::default(),
-        conn.clone(),
-        SQLiteTransactionStore::new(conn.clone()),
-    );
+    let app_config = match AppState::new(conn, &secret, Default::default()) {
+        Ok(config) => config,
+        Err(error) => {
+            eprintln!("Could not initialize database: {error}");
+            exit(1);
+        }
+    };
 
     let handle = Handle::new();
     tokio::spawn(graceful_shutdown(handle.clone()));
