+++ src/db.rs
@@ -4,7 +4,7 @@ use rusqlite::{Connection, Error, Row, Transaction as SqlTransaction};
 
 use crate::{
     balances::create_balance_table, category::create_category_table,
-    stores::sqlite::SQLiteTransactionStore, user::create_user_table,
+    transaction::create_transaction_table, user::create_user_table,
 };
 
 /// A trait for adding an object schema to a database.
@@ -130,7 +130,7 @@ pub fn initialize(connection: &Connection) -> Result<(), Error> {
     create_user_table(&transaction)?;
     create_balance_table(&transaction)?;
     create_category_table(&transaction)?;
-    SQLiteTransactionStore::create_table(&transaction)?;
+    create_transaction_table(&transaction)?;
 
     transaction.commit()?;
