+++ src/routes/views/dashboard.rs
@@ -13,7 +13,7 @@ use crate::{
         navigation::{NavbarTemplate, get_nav_bar},
     },
     state::DashboardState,
-    stores::{TransactionQuery, TransactionStore},
+    transaction::{TransactionQuery, query_transactions},
 };
 
 /// Renders the dashboard page.
@@ -26,10 +26,7 @@ struct DashboardTemplate<'a> {
 }
 
 /// Display a page with an overview of the user's data.
-pub async fn get_dashboard_page<T>(State(state): State<DashboardState<T>>) -> Response
-where
-    T: TransactionStore + Send + Sync,
-{
+pub async fn get_dashboard_page(State(state): State<DashboardState>) -> Response {
     let nav_bar = get_nav_bar(endpoints::DASHBOARD_VIEW);
 
     let today = OffsetDateTime::now_utc().date();
@@ -44,10 +41,14 @@ where
         }
     };
 
-    let transactions = state.transaction_store.get_query(TransactionQuery {
-        date_range: Some(one_week_ago..=today),
-        ..Default::default()
-    });
+    let connection = state.db_connection.lock().unwrap();
+    let transactions = query_transactions(
+        TransactionQuery {
+            date_range: Some(one_week_ago..=today),
+            ..Default::default()
+        },
+        &connection,
+    );
 
     let balance = match transactions {
         Ok(transactions) => transactions
@@ -70,77 +71,26 @@ mod dashboard_route_tests {
     use time::{Duration, OffsetDateTime};
 
     use crate::{
-        Error,
-        models::{DatabaseID, Transaction, TransactionBuilder},
-        state::DashboardState,
-        stores::{TransactionQuery, TransactionStore},
+        db::initialize, models::Transaction, state::DashboardState, transaction::create_transaction,
     };
+    use rusqlite::Connection;
+    use std::sync::{Arc, Mutex};
 
     use super::get_dashboard_page;
 
-    #[derive(Clone)]
-    struct FakeTransactionStore {
-        transactions: Vec<Transaction>,
-    }
-
-    impl TransactionStore for FakeTransactionStore {
-        fn create(&mut self, amount: f64) -> Result<Transaction, Error> {
-            self.create_from_builder(TransactionBuilder::new(amount))
-        }
-
-        fn create_from_builder(
-            &mut self,
-            builder: TransactionBuilder,
-        ) -> Result<Transaction, Error> {
-            let next_id = match self.transactions.last() {
-                Some(transaction) => transaction.id() + 1,
-                None => 0,
-            };
-
-            let transaction = builder.finalise(next_id);
-
-            self.transactions.push(transaction.clone());
-
-            Ok(transaction)
-        }
-
-        fn import(
-            &mut self,
-            _builders: Vec<TransactionBuilder>,
-        ) -> Result<Vec<Transaction>, Error> {
-            todo!()
-        }
-
-        fn get(&self, _id: DatabaseID) -> Result<Transaction, Error> {
-            todo!()
-        }
-
-        fn get_query(&self, filter: TransactionQuery) -> Result<Vec<Transaction>, Error> {
-            self.transactions
-                .iter()
-                .filter(|transaction| {
-                    let mut should_keep = true;
-
-                    if let Some(ref date_range) = filter.date_range {
-                        should_keep &= date_range.start() <= transaction.date()
-                            && transaction.date() <= date_range.end();
-                    }
-
-                    should_keep
-                })
-                .map(|transaction| Ok(transaction.to_owned()))
-                .collect()
-        }
-
-        fn count(&self) -> Result<usize, Error> {
-            todo!()
-        }
+    fn get_test_connection() -> Connection {
+        let conn = Connection::open_in_memory().unwrap();
+        initialize(&conn).unwrap();
+        conn
     }
 
     #[tokio::test]
     async fn dashboard_displays_correct_balance() {
-        let transactions = vec![
-            // Transactions before the current week should not be included in the balance.
+        let conn = get_test_connection();
+
+        // Create transactions in the database
+        // Transaction before the current week should not be included in the balance
+        create_transaction(
             Transaction::build(12.3)
                 .date(
                     OffsetDateTime::now_utc()
@@ -148,15 +98,18 @@ mod dashboard_route_tests {
                         .checked_sub(Duration::weeks(2))
                         .unwrap(),
                 )
-                .unwrap()
-                .finalise(1),
-            // These transactions should be included.
-            Transaction::build(45.6).finalise(2),
-            Transaction::build(-45.6).finalise(3),
-            Transaction::build(123.0).finalise(4),
-        ];
+                .unwrap(),
+            &conn,
+        )
+        .unwrap();
+
+        // These transactions should be included
+        create_transaction(Transaction::build(45.6), &conn).unwrap();
+        create_transaction(Transaction::build(-45.6), &conn).unwrap();
+        create_transaction(Transaction::build(123.0), &conn).unwrap();
+
         let state = DashboardState {
-            transaction_store: FakeTransactionStore { transactions },
+            db_connection: Arc::new(Mutex::new(conn)),
         };
 
         let response = get_dashboard_page(State(state)).await;
@@ -167,9 +120,13 @@ mod dashboard_route_tests {
 
     #[tokio::test]
     async fn dashboard_displays_negative_balance_without_sign() {
-        let transactions = vec![Transaction::build(-123.0).finalise(2)];
+        let conn = get_test_connection();
+
+        // Create transaction in the database
+        create_transaction(Transaction::build(-123.0), &conn).unwrap();
+
         let state = DashboardState {
-            transaction_store: FakeTransactionStore { transactions },
+            db_connection: Arc::new(Mutex::new(conn)),
         };
 
         let response = get_dashboard_page(State(state)).await;
