--- src/stores/mod.rs
@@ -1,7 +1,0 @@
-//! Contains traits and implementations for objects that store the domain [models](crate::models).
-
-mod transaction;
-
-pub mod sqlite;
-
-pub use transaction::{SortOrder, TransactionQuery, TransactionStore};
